refactor: extract SessionLibraryQuery.swift (#665 slice B)
Second slice of #665. Extracts 4 query DSL types:
- enum SessionLibraryDateFilter
- enum SessionLibrarySortOrder
- struct SessionLibraryDateRange
- struct SessionLibraryQuery

Byte-identical move; SHA-256 e0e32b6e66b0408a493fc2f46c0a8cb710c7724349c1bc66283445e683b53258
verified against origin/main:SessionLibrary.swift lines 3-61 (post-#665 slice A).

Reduces SessionLibrary.swift from 327 to 268 lines.

Tests: 667 (unchanged).
